Move tests to a single source
